+++ src/IfcExpressionOptions.ts
@@ -1,0 +1,5 @@
+import { BuiltinVariableRegistry } from "./builtin/BuiltinVariableRegistry.js";
+
+export interface IfcExpressionOptions {
+  builtinVariableRegistry?: BuiltinVariableRegistry;
+}