+++ src/compiler/ExprCompiler.ts
@@ -11,6 +11,8 @@ import {
   LogicalLiteralContext,
   MethodCallChainEndContext,
   MethodCallChainInnerContext,
+  MethodFunctionCallContext,
+  MethodPropertyAccessContext,
   NumLiteralContext,
   SEAddSubContext,
   SEArrayExprContext,
@@ -37,8 +39,6 @@ import { MinusExpr } from "../expression/numeric/MinusExpr.js";
 import { ParserRuleContext } from "antlr4";
 import { MultiplyExpr } from "../expression/numeric/MultiplyExpr.js";
 import { DivideExpr } from "../expression/numeric/DivideExpr.js";
-import { PropObjectReferenceExpr } from "../expression/reference/PropObjectReferenceExpr.js";
-import { ElemObjectReferenceExpr } from "../expression/reference/ElemObjectReferenceExpr.js";
 import { StringLiteralExpr } from "../expression/string/StringLiteralExpr.js";
 import { ArrayExpr } from "../expression/structure/ArrayExpr.js";
 import { FunctionExpr } from "../expression/function/FunctionExpr.js";
@@ -64,16 +64,29 @@ import { LessThanOrEqual } from "../expression/comparison/LessThanOrEqual.js";
 import { ExprManager } from "./ExprManager.js";
 import { LogicalValue } from "../value/LogicalValue.js";
 import { LogicalLiteralExpr } from "../expression/boolean/LogicalLiteralExpr.js";
+import {
+  BuiltinVariableRegistry,
+  isBuiltinFunctionDefinition,
+  isBuiltinPropertyDefinition,
+} from "../builtin/BuiltinVariableRegistry.js";
+import { ContextObjectType } from "../type/ContextObjectType.js";
+import { BuiltinPropertyAccessExpr } from "../expression/reference/BuiltinPropertyAccessExpr.js";
+import { BuiltinFunctionCallExpr } from "../expression/reference/BuiltinFunctionCallExpr.js";
 
 export class ExprCompiler extends IfcExpressionVisitor<Expr<any>> {
-  private readonly methodCallTargetStack = [];
+  private readonly methodCallTargetStack: Array<Expr<any>> = [];
   private readonly typeManager: TypeManager;
   private readonly exprManager: ExprManager;
+  private readonly builtinVariableRegistry: BuiltinVariableRegistry;
 
-  constructor(typeManager: TypeManager) {
+  constructor(
+    typeManager: TypeManager,
+    builtinVariableRegistry: BuiltinVariableRegistry = BuiltinVariableRegistry.getDefaultRegistry()
+  ) {
     super();
     this.typeManager = typeManager;
     this.exprManager = new ExprManager();
+    this.builtinVariableRegistry = builtinVariableRegistry;
   }
 
   public getExprManager(): ExprManager {
@@ -84,20 +97,17 @@ export class ExprCompiler extends IfcExpressionVisitor<Expr<any>> {
     return this.visit(ctx.singleExpr());
   };
 
-  /*================================================
-   * VariableRef
-   *==============================================*/
-
   private associateContextAndReturn(expr: Expr<any>, ctx: ParserRuleContext) {
     this.exprManager.registerExprWithContext(expr, ctx);
     return expr;
   }
 
   visitVariableRef: (ctx: VariableRefContext) => Expr<any> = (ctx) => {
-    if (ctx.IDENTIFIER().getText().toUpperCase() === "PROPERTY") {
-      return this.associateContextAndReturn(new PropObjectReferenceExpr(), ctx);
-    } else if (ctx.IDENTIFIER().getText().toUpperCase() === "ELEMENT") {
-      return this.associateContextAndReturn(new ElemObjectReferenceExpr(), ctx);
+    const builtin = this.builtinVariableRegistry.getDefinition(
+      ctx.IDENTIFIER().getText()
+    );
+    if (!isNullish(builtin)) {
+      return this.associateContextAndReturn(builtin.createReferenceExpr(), ctx);
     }
     throw new Error(`Parsing error: No variable '${ctx.getText()}' found  `);
   };
@@ -163,35 +173,71 @@ export class ExprCompiler extends IfcExpressionVisitor<Expr<any>> {
     );
   };
 
-  /*================================================
-   * MethodCallChain
-   *==============================================*/
-
   visitMethodCallChainInner: (ctx: MethodCallChainInnerContext) => Expr<any> = (
     ctx
   ) => {
-    const functionExpr = new FunctionExpr(
-      ctx.functionCall().IDENTIFIER().getText(),
-      this.collectFunctionArguments(ctx.functionCall().exprList(), [
-        this.methodCallTargetStack.pop(),
-      ])
-    );
-    this.associateContextAndReturn(functionExpr, ctx.functionCall());
-    this.methodCallTargetStack.push(functionExpr);
-    return this.visit(ctx.methodCallChain());
+    const accessExpr = this.visit(ctx._target);
+    this.methodCallTargetStack.push(accessExpr);
+    return this.visit(ctx._call);
   };
 
   visitMethodCallChainEnd: (ctx: MethodCallChainEndContext) => Expr<any> = (
     ctx
   ) => {
+    return this.visit(ctx._call);
+  };
+
+  visitMethodFunctionCall: (ctx: MethodFunctionCallContext) => Expr<any> = (
+    ctx
+  ) => {
+    const targetExpr = this.methodCallTargetStack.pop();
+    const functionName = ctx.functionCall().IDENTIFIER().getText();
+    const args = isNullish(ctx.functionCall().exprList())
+      ? []
+      : this.collectFunctionArguments(ctx.functionCall().exprList());
+    const targetType = targetExpr.getType();
+    if (targetType instanceof ContextObjectType) {
+      const member = targetType.getMemberDefinition(functionName);
+      if (isBuiltinFunctionDefinition(member)) {
+        return this.associateContextAndReturn(
+          new BuiltinFunctionCallExpr(
+            targetExpr,
+            member.name,
+            args,
+            member.returnType,
+            targetType.getName()
+          ),
+          ctx
+        );
+      }
+    }
     return this.associateContextAndReturn(
-      new FunctionExpr(
-        ctx.functionCall().IDENTIFIER().getText(),
-        this.collectFunctionArguments(ctx.functionCall().exprList(), [
-          this.methodCallTargetStack.pop(),
-        ])
-      ),
+      new FunctionExpr(functionName, [targetExpr, ...args]),
       ctx.functionCall()
+    );
+  };
+
+  visitMethodPropertyAccess: (ctx: MethodPropertyAccessContext) => Expr<any> = (
+    ctx
+  ) => {
+    const targetExpr = this.methodCallTargetStack.pop();
+    const targetType = targetExpr.getType();
+    if (targetType instanceof ContextObjectType) {
+      const member = targetType.getMemberDefinition(ctx.IDENTIFIER().getText());
+      if (isBuiltinPropertyDefinition(member)) {
+        return this.associateContextAndReturn(
+          new BuiltinPropertyAccessExpr(
+            targetExpr,
+            member.name,
+            member.valueType,
+            targetType.getName()
+          ),
+          ctx
+        );
+      }
+    }
+    throw new Error(
+      `Parsing error: No property '${ctx.getText()}' found for type '${targetType.getName()}'`
     );
   };
 
@@ -261,19 +307,11 @@ export class ExprCompiler extends IfcExpressionVisitor<Expr<any>> {
     }
   };
 
-  /*================================================
-   * StringExpr
-   *===============================================*/
-
   visitStringLiteral: (ctx: StringLiteralContext) => Expr<any> = (ctx) => {
     const quotedString = ctx.QUOTED_STRING().getText();
     const text = quotedString.substring(1, quotedString.length - 1);
     return this.associateContextAndReturn(new StringLiteralExpr(text), ctx);
   };
-
-  /*================================================
-   * NumExpr
-   *===============================================*/
 
   visitNumLiteral: (ctx: NumLiteralContext) => Expr<any> = (ctx) => {
     let val = ctx.INT();
@@ -310,10 +348,6 @@ export class ExprCompiler extends IfcExpressionVisitor<Expr<any>> {
       ctx
     );
   };
-
-  /*================================================
-   * BooleanExpr
-   *===============================================*/
 
   visitSEBooleanBinaryOp: (ctx: SEBooleanBinaryOpContext) => Expr<any> = (
     ctx
@@ -360,7 +394,6 @@ export class ExprCompiler extends IfcExpressionVisitor<Expr<any>> {
   };
 
   visitLogicalLiteral: (ctx: LogicalLiteralContext) => Expr<any> = (ctx) => {
-    // the only literal we recognize is 'UNKNOWN' (true and false are boolean literals)
     return this.associateContextAndReturn(
       new LogicalLiteralExpr(LogicalValue.UNKNOWN_VALUE),
       ctx
@@ -373,10 +406,6 @@ export class ExprCompiler extends IfcExpressionVisitor<Expr<any>> {
       ctx
     );
   };
-
-  /*================================================
-   * ArrayExpr
-   *===============================================*/
 
   visitArrayExpr: (ctx: ArrayExprContext) => Expr<any> = (ctx) => {
     return this.associateContextAndReturn(
@@ -400,10 +429,6 @@ export class ExprCompiler extends IfcExpressionVisitor<Expr<any>> {
     }
     return [];
   };
-
-  /*================================================
-   * FuncExpr
-   *===============================================*/
 
   visitFunctionCall: (ctx: FunctionCallContext) => Expr<any> = (ctx) => {
     const args = isNullish(ctx.exprList())
@@ -440,3 +465,11 @@ export class ExprCompiler extends IfcExpressionVisitor<Expr<any>> {
     return this.visit(ctx.getChild(0));
   };
 }
+
+
+
+
+
+
+
+